--- addon/locales/en/components/fd-settings-appstruct.js
@@ -1,8 +1,0 @@
-export default {
-  form: 'The form',
-  formName: 'Form class name',
-  formType: 'Type (stereotype)',
-  title: 'Navigation element title',
-  url: 'Url',
-  description: 'Description'
-};
