--- addon/locales/ru/components/fd-settings-appstruct.js
@@ -1,8 +1,0 @@
-export default {
-  form: 'Форма',
-  formName: 'Название класса формы',
-  formType: 'Тип (стереотип)',
-  title: 'Заголовок элемента навигации',
-  url: 'Ссылка',
-  description: 'Описание'
-};
