--- app/components/fd-settings-appstruct.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-designer/components/fd-settings-appstruct';
